+++ src/tools/task-tools.ts
@@ -382,6 +382,10 @@ export const createSubtaskTool: Tool = {
         type: "string",
         description: "Description of the subtask"
       },
+      html_notes: {
+        type: "string",
+        description: "HTML-like formatted description of the subtask. Does not support ALL HTML tags. Only a subset. The only allowed TAG in the HTML are: <body> <h1> <h2> <ol> <ul> <li> <strong> <em> <u> <s> <code> <pre> <blockquote> <a data-asana-type=\"\" data-asana-gid=\"\"> <hr> <img> <table> <tr> <td>. No other tags are allowed. Use the \\n to create a newline. Do not use \\n after <body>. Example: <body><h1>Motivation</h1>\nA customer called in to complain\n<h1>Goal</h1>\nFix the problem</body>"
+      },
       due_on: {
         type: "string",
         description: "Due date in YYYY-MM-DD format"